Add tools menu to clear old dots, edit text, open student view
Replaces the "Clear Old Markers" button with a button to open a little
toolbox for those features. Clicking on the rest of the page (or the
tools button again) closes the toolbox, so it should be relatively
user-friendly.

The tools icon is in the public domain, from the Tango Desktop Project.